*** empty log message ***
svn path=/trunk/boinc/; revision=12249

+++ clientgui/DlgAdvPreferences.cpp
@@ -181,9 +181,9 @@ void CDlgAdvPreferences::ReadPreferenceSettings() {
 	// ######### proc usage page
 	// do work between
 	m_rbtProcEveryDay->SetValue(true);
-	buffer.Printf(wxT("%02d:00"),prefs.time_prefs.start_hour);
+	buffer.Printf(wxT("%02d:00"),(int)prefs.time_prefs.start_hour);
 	*m_txtProcEveryDayStart << buffer;
-	buffer.Printf(wxT("%02d:00"),prefs.time_prefs.end_hour);
+	buffer.Printf(wxT("%02d:00"),(int)prefs.time_prefs.end_hour);
 	*m_txtProcEveryDayStop << buffer;
 	// on batteries
 	m_chkProcOnBatteries->SetValue(prefs.run_on_batteries);
@@ -205,9 +205,9 @@ void CDlgAdvPreferences::ReadPreferenceSettings() {
 	// ######### net usage page
 	// use network between
 	m_rbtNetEveryDay->SetValue(true);
-	buffer.Printf(wxT("%02d:00"),prefs.time_prefs.net_start_hour);
+	buffer.Printf(wxT("%02d:00"),(int)prefs.time_prefs.net_start_hour);
 	*m_txtNetEveryDayStart << buffer;
-	buffer.Printf(wxT("%02d:00"),prefs.time_prefs.net_end_hour);
+	buffer.Printf(wxT("%02d:00"),(int)prefs.time_prefs.net_end_hour);
 	*m_txtNetEveryDayStop << buffer;
 	// connection interval
 	buffer.Printf(wxT("%01.4f"),prefs.work_buf_min_days);
